feat: added more locales (#50)
* feat: add preParse and PostFormat to locale definitions

Add locale 'ar' as an example

* feat: add all locales from dayjs

* test: add tests for all added locales

* docs: add documentation for supported locales

* refactor: remove usage of 'instanceof' in plugins/locale/index.ts

* docs: update comments for yearStart

yearStart = 7 + weekStart - janX
janX is first day of january that must belong to first week of year

* feat: make en-us the default locale (like momentjs)

* fix: replace value of locale.yearStart with day number

The current value was the "doy" value from moment, but esday uses a
different algorithm and requires the day number.

* test: update tests for locales to match updated yearStart values

* fix: weekOfYear fails for 1st day of week after start of DST

* refactor: fix biome findings

* refactor: rename DATE to DATE_OF_WEEK

I always was looking if day ord date is day of month

* fix: when crossing the DST border in a diff utcOffset was wrong

* fix: make WeekOfYear return the correct value on the first day of a week

* test: update import path for util.ts

* chore: suppress screenshots in browser tests

* test: work-around to prevent errors in vitest browser tests

In vitest browser tests momentjs cannot load the locales (as it does
not find the locale file). As a work-around the 2 test files using
locales are excluded from running in the browser